Added socket connection/client handler
Each client has a window of 5 seconds to send a valid message to the window manager upon connecting to its socket.
If the client does not do so, it is disconnected. Otherwise, its socket's file descriptor is added to a list of file descriptors to monitor.
Any message sent from the client, on that socket, is received and handled by the window manager.