Switched error and diagnostics logging to corresponding macros
The macros FWM_ELOG and FWM_DLOG include the name of the file and function, and the line number where they're invoked.
This is useful for debugging, but inconvenient to implement outside a macro.
The macros, however, make use of `##__VA_ARGS__` to preserve the variadic nature of the `fwm_log` function, which they call.
`##__VA_ARGS__` is a GNU extension, and ideally should not be used.